updated functionality

+++ src/main/java/com/imagegrafia/petrolpump/repository/PumpRepository.java
@@ -9,7 +9,7 @@ import com.imagegrafia.petrolpump.entity.Nozzle;
 import com.imagegrafia.petrolpump.entity.Pump;
 
 @Repository
-public interface PumpRepository extends CrudRepository<Pump, Long>{
+public interface PumpRepository extends CrudRepository<Pump, Integer>{
